requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
pyodbc==5.1.0
python-dotenv==1.0.1


#добавить 2 аналитических запроса и сложный с таблицами

В четрверг в 11
Если будут замечания, то договоримся на пятницу